requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
cloudinary==1.36.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
executing==1.2.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
Markdown==3.4.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
supervisor==4.2.5
tomli==2.0.1
traitlets==5.9.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
zope.interface==6.0
